+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add support for scrollable views. Activate via the option `scrollable: true`. Once you activate scrollable views all views are automatically zoomfixed! See [http://localhost:8080/others/scrollable-container.html](others/scrollable-container.html) for an example.
 - Add `option(key, value)` to the JS API for changing options. It supports getting all options when `value` is ommited and setting `scrollable`.
 - Properly display "Loading" while loading tileset info
+- Add support to ignore offscreen values for value scale locking by setting `ignoreOffScreenValues: true` for a lock group in `locksDict`.
 
 _[Detailed changes since v1.6.12](https://github.com/higlass/higlass/compare/v1.6.12...v1.6.13)_